--- src/app/modules/dashboard/dashboard.module.ts
@@ -1,45 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { DashboardComponent } from './dashboard.component';
-import { SearchComponent } from './search/search.component';
-import { HeaderComponent } from './header/header.component';
-import { FooterComponent } from './footer/footer.component';
-import { FormsModule,ReactiveFormsModule } from '@angular/forms';
-
-import { NodeService } from './dashboardservice';
-import {TreeTableModule} from 'primeng/treetable';
-import {ButtonModule} from 'primeng/button';
-import {DialogModule} from 'primeng/dialog';
-import {MultiSelectModule} from 'primeng/multiselect';
-import {InputTextModule} from 'primeng/inputtext';
-import {ToastModule} from 'primeng/toast';
-import {ContextMenuModule} from 'primeng/contextmenu';
-import {TableModule} from 'primeng/table';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { CollectionComponent } from './collection/collection.component';
-
-
-
-@NgModule({
-  declarations: [DashboardComponent, SearchComponent, HeaderComponent, FooterComponent, CollectionComponent],
-  imports: [
-    CommonModule,
-     TreeTableModule,
-     BrowserAnimationsModule,
-     TableModule,
-    ToastModule,
-    DialogModule,
-    ButtonModule,
-    MultiSelectModule,
-    InputTextModule,
-    ContextMenuModule,
-    FormsModule,
-    ReactiveFormsModule,
-  ],
-  exports: [
-    DashboardComponent,
-    SearchComponent
-  ],
-  providers: [NodeService]
-})
-export class DashboardModule { }
