--- src/app/modules/dashboard/search/data.ts
@@ -1,4627 +1,0 @@
-/**
- * Test cases data source
- */
-export let sampleData: Object[] = [
-   
-          {
-            "bibId": 1993640,
-            "title": "The ... Agricultural survey.",
-            "author": "   ",
-            "publisher": "The Dept. of Agriculture,",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1967/68-1972/73; 1977",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "9434.202q",
-                "chronologyAndEnum": "1967/68",
-                "customerCode": "PA",
-                "barcode": "32101043700374",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 3043120,
-                "owningInstitutionItemId": "3131946",
-                "owningInstitutionHoldingsId": "953209"
-              },
-              {
-                "callNumber": "9434.202q",
-                "chronologyAndEnum": "1969/70",
-                "customerCode": "PA",
-                "barcode": "32101059090363",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 3043121,
-                "owningInstitutionItemId": "4304289",
-                "owningInstitutionHoldingsId": "953209"
-              },
-              {
-                "callNumber": "9434.202q",
-                "chronologyAndEnum": "1970/71",
-                "customerCode": "PA",
-                "barcode": "32101059090371",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 3043122,
-                "owningInstitutionItemId": "4304292",
-                "owningInstitutionHoldingsId": "953209"
-              },
-              {
-                "callNumber": "9434.202q",
-                "chronologyAndEnum": "1971/72",
-                "customerCode": "PA",
-                "barcode": "32101059090389",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 3043123,
-                "owningInstitutionItemId": "4304294",
-                "owningInstitutionHoldingsId": "953209"
-              },
-              {
-                "callNumber": "9434.202q",
-                "chronologyAndEnum": "1972/73",
-                "customerCode": "PA",
-                "barcode": "32101059090397",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 3043124,
-                "owningInstitutionItemId": "4304296",
-                "owningInstitutionHoldingsId": "953209"
-              },
-              {
-                "callNumber": "9434.202q",
-                "chronologyAndEnum": "1977",
-                "customerCode": "PA",
-                "barcode": "32101059090405",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 3043125,
-                "owningInstitutionItemId": "4304300",
-                "owningInstitutionHoldingsId": "953209"
-              }
-            ],
-            "owningInstitutionBibId": "837187",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 8891626,
-            "title": "... ajandası / Kadın Eserleri Kütüphanesi ve Bilgi Merkezi Vakfı.",
-            "author": "   ",
-            "publisher": "Kadın Eserleri Kütüphanesi ve Bilgi Merkezi Vakfı,",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "2016-2017",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HQ1726.7 .A326",
-                "chronologyAndEnum": "2016",
-                "customerCode": "PA",
-                "barcode": "32101101289450",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14681098,
-                "owningInstitutionItemId": "7731992",
-                "owningInstitutionHoldingsId": "10256961"
-              },
-              {
-                "callNumber": "HQ1726.7 .A326",
-                "chronologyAndEnum": "2017",
-                "customerCode": "PA",
-                "barcode": "32101101289443",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14681077,
-                "owningInstitutionItemId": "7731993",
-                "owningInstitutionHoldingsId": "10256961"
-              }
-            ],
-            "owningInstitutionBibId": "10503953",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1579711,
-            "title": "[ Analyses of Hour and Wage Provisions of Codes Approved].",
-            "author": " United States. Women's Bureau.  ",
-            "publisher": "Women's Bureau,",
-            "publisherDate": "1933].",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "No. 1159 (Aug., 1933) - no. 1356 (June, 1934)",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "IR-12207",
-                "chronologyAndEnum": "no.1159-1340 (Aug., 1933 - April, 1934)",
-                "customerCode": "PA",
-                "barcode": "32101077419008",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2224561,
-                "owningInstitutionItemId": "5938134",
-                "owningInstitutionHoldingsId": "6374676"
-              },
-              {
-                "callNumber": "IR-12207",
-                "chronologyAndEnum": "no.1179-1254 (Aug., 1933 - Dec., 1933)",
-                "customerCode": "PA",
-                "barcode": "32101077419016",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2224562,
-                "owningInstitutionItemId": "5938165",
-                "owningInstitutionHoldingsId": "6374676"
-              },
-              {
-                "callNumber": "IR-12207",
-                "chronologyAndEnum": "no.1257-1356 (Dec., 1933 - June, 1934)",
-                "customerCode": "PA",
-                "barcode": "32101077419024",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2224563,
-                "owningInstitutionItemId": "5938197",
-                "owningInstitutionHoldingsId": "6374676"
-              }
-            ],
-            "owningInstitutionBibId": "6386321",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1934216,
-            "title": "... Anglo-American Aeronautical Conference.",
-            "author": "  Anglo-American Aeronautical Conference. ",
-            "publisher": "The Royal Aeronautical Society,",
-            "publisherDate": "1952.",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "3rd (4th-7th Sept. 1951)-4th (15th-17th Sept. 1953)",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "9250.114",
-                "chronologyAndEnum": "1951",
-                "customerCode": "PA",
-                "barcode": "32101045244348",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2930065,
-                "owningInstitutionItemId": "3925537",
-                "owningInstitutionHoldingsId": "782633"
-              },
-              {
-                "callNumber": "9250.114",
-                "chronologyAndEnum": "1953",
-                "customerCode": "PA",
-                "barcode": "32101045244330",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2930066,
-                "owningInstitutionItemId": "3925540",
-                "owningInstitutionHoldingsId": "782633"
-              }
-            ],
-            "owningInstitutionBibId": "701147",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1935273,
-            "title": "... Anglo-American Aeronautical Conference.",
-            "author": "  Anglo-American Aeronautical Conference. ",
-            "publisher": "Royal Aeronautical Society,",
-            "publisherDate": "1959-",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "6th (9th-12th Sept. 1957)-9th (Oct. 16-24, 1963); 15th (31 May-2 June 1977)",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "9250.114",
-                "chronologyAndEnum": "15th (1977)",
-                "customerCode": "PA",
-                "barcode": "32101045244512",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2944252,
-                "owningInstitutionItemId": "3925578",
-                "owningInstitutionHoldingsId": "782672"
-              },
-              {
-                "callNumber": "9250.114",
-                "chronologyAndEnum": "6th , (1957)",
-                "customerCode": "PA",
-                "barcode": "32101045244454",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2944248,
-                "owningInstitutionItemId": "3925561",
-                "owningInstitutionHoldingsId": "782672"
-              },
-              {
-                "callNumber": "9250.114",
-                "chronologyAndEnum": "7th, (1959)",
-                "customerCode": "PA",
-                "barcode": "32101045244447",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2944249,
-                "owningInstitutionItemId": "3925568",
-                "owningInstitutionHoldingsId": "782672"
-              },
-              {
-                "callNumber": "9250.114",
-                "chronologyAndEnum": "8th (1961)",
-                "customerCode": "PA",
-                "barcode": "32101045244439",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2944250,
-                "owningInstitutionItemId": "3925574",
-                "owningInstitutionHoldingsId": "782672"
-              },
-              {
-                "callNumber": "9250.114",
-                "chronologyAndEnum": "9th (1963)",
-                "customerCode": "PA",
-                "barcode": "32101045244421",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2944251,
-                "owningInstitutionItemId": "3925576",
-                "owningInstitutionHoldingsId": "782672"
-              }
-            ],
-            "owningInstitutionBibId": "701181",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1877691,
-            "title": "... Annual Conference and General Meeting.",
-            "author": " East African Law Society. Annual Conference and General Meeting, author.  ",
-            "publisher": "East African Law Society",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "20th (2015)-21st (2016)",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "KQC74.A353 .E197",
-                "chronologyAndEnum": "20th (2015)",
-                "customerCode": "PA",
-                "barcode": "32101098726506",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2792084,
-                "owningInstitutionItemId": "7583399",
-                "owningInstitutionHoldingsId": "9919989"
-              },
-              {
-                "callNumber": "KQC74.A353 .E197",
-                "chronologyAndEnum": "21st (2016)",
-                "customerCode": "PA",
-                "barcode": "32101097391245",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14585586,
-                "owningInstitutionItemId": "7703963",
-                "owningInstitutionHoldingsId": "9919989"
-              }
-            ],
-            "owningInstitutionBibId": "10128754",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 530453,
-            "title": "... Annual conference of state and local boards of health.",
-            "author": " New Jersey. Board of Health.  ",
-            "publisher": "Board of Health of the State of New Jersey,",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1907",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "8941.674.2",
-                "chronologyAndEnum": "1907",
-                "customerCode": "PA",
-                "barcode": "32101058089093",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 644680,
-                "owningInstitutionItemId": "4214289",
-                "owningInstitutionHoldingsId": "2461067"
-              }
-            ],
-            "owningInstitutionBibId": "2174942",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 9228836,
-            "title": "... annual report.",
-            "author": " Southern Sudan Human Rights Commission, author,  ",
-            "publisher": "Southern Sudan Human Rights Commission,",
-            "publisherDate": "[2008]-",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1st (July 2006/December 2007)-2nd (2008)",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "JC599.S643 S68",
-                "chronologyAndEnum": "1st (July/Dec. 2007)",
-                "customerCode": "PA",
-                "barcode": "32101103736482",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 15241815,
-                "owningInstitutionItemId": "8009717",
-                "owningInstitutionHoldingsId": "10845745"
-              },
-              {
-                "callNumber": "JC599.S643 S68",
-                "chronologyAndEnum": "2nd (2008)",
-                "customerCode": "PA",
-                "barcode": "32101103736490",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 15241874,
-                "owningInstitutionItemId": "8009719",
-                "owningInstitutionHoldingsId": "10845745"
-              }
-            ],
-            "owningInstitutionBibId": "11131831",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1823309,
-            "title": "... annual report.",
-            "author": " Chamber of Mines (Zimbabwe), author,  ",
-            "publisher": "The Chamber of Mines of Zimbabwe",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "2013-2014",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "TN119.R45 C46aq",
-                "chronologyAndEnum": "2013",
-                "customerCode": "PA",
-                "barcode": "32101092836939",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2725149,
-                "owningInstitutionItemId": "7192752",
-                "owningInstitutionHoldingsId": "9027500"
-              },
-              {
-                "callNumber": "TN119.R45 C46aq",
-                "chronologyAndEnum": "2014",
-                "customerCode": "PA",
-                "barcode": "32101099156919",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2725150,
-                "owningInstitutionItemId": "7463298",
-                "owningInstitutionHoldingsId": "9027500"
-              }
-            ],
-            "owningInstitutionBibId": "9143559",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 562146,
-            "title": "... Annual report / Division of the New Jersey Real Estate Commission in the Department of Banking and Insurance.",
-            "author": " New Jersey. Division of the New Jersey Real Estate Commission.  ",
-            "publisher": "The Commission,",
-            "publisherDate": "1954-",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1953/54-  1957/58 -1959/60",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HD266.N4A5",
-                "chronologyAndEnum": "1957/58-1959/1960",
-                "customerCode": "PA",
-                "barcode": "32101047854060",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 682551,
-                "owningInstitutionItemId": "2892927",
-                "owningInstitutionHoldingsId": "2515920"
-              }
-            ],
-            "owningInstitutionBibId": "2228806",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 9419892,
-            "title": "... annual report / NBFIRA, Non-Bank Financial Institutions Regulatory Authority.",
-            "author": " Non-Bank Financial Institutions Regulatory Authority (Botswana), author,  ",
-            "publisher": "Non-Bank Financial Institutions Regulatory Authority",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "2014",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HG187.5.B55 N66aq",
-                "chronologyAndEnum": "2014",
-                "customerCode": "PA",
-                "barcode": "32101107802785",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 15645570,
-                "owningInstitutionItemId": "8146249",
-                "owningInstitutionHoldingsId": "11160294"
-              }
-            ],
-            "owningInstitutionBibId": "11471029",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1934061,
-            "title": "... Annual report of Board of Directors and balance sheet, December 31 ...",
-            "author": " Amerada Corporation.  ",
-            "publisher": "The Corporation,",
-            "publisherDate": "1926-[1941]",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "5th (1925)-[20th] (1940)",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HD9569.A5 A3",
-                "chronologyAndEnum": "1925/1945",
-                "customerCode": "PA",
-                "barcode": "32101089725491",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2928425,
-                "owningInstitutionItemId": "6939747",
-                "owningInstitutionHoldingsId": "779054"
-              }
-            ],
-            "owningInstitutionBibId": "697999",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 760631,
-            "title": "... Annual report of the directors of the Northern Railroad to the stockholders.",
-            "author": " Northern Railroad Company (N.H.)  ",
-            "publisher": "s.n.,",
-            "publisherDate": "1846-",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "27th-47th",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HE2791.N83A1",
-                "chronologyAndEnum": "27th-47th (1871/72-1891/92)",
-                "customerCode": "PA",
-                "barcode": "32101066786888",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 990515,
-                "owningInstitutionItemId": "4944825",
-                "owningInstitutionHoldingsId": "2784446"
-              }
-            ],
-            "owningInstitutionBibId": "2483208",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 677547,
-            "title": "... Annual report of the managers and officers of the New Jersey State Hospital at Greystone Park : for the year ending ...",
-            "author": " New Jersey State Hospital at Greystone Park.  ",
-            "publisher": "The Hospital",
-            "publisherDate": "1925-1959",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1943/44-1948/49, 1951/52-1959",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HV3006.N3G86",
-                "chronologyAndEnum": "1943/44-1948/49 (inc.)",
-                "customerCode": "PA",
-                "barcode": "32101057576876",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 851126,
-                "owningInstitutionItemId": "4260193",
-                "owningInstitutionHoldingsId": "2640265"
-              },
-              {
-                "callNumber": "HV3006.N3G86",
-                "chronologyAndEnum": "1951/52-1960/61",
-                "customerCode": "PA",
-                "barcode": "32101057576884",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 851125,
-                "owningInstitutionItemId": "4260192",
-                "owningInstitutionHoldingsId": "2640265"
-              }
-            ],
-            "owningInstitutionBibId": "2347124",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1468728,
-            "title": "... Annual report of the president, treasurer and attorney ... for the year ...",
-            "author": "   ",
-            "publisher": null,
-            "publisherDate": "[19--]",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "47th annual, 1922",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "IR-7973",
-                "chronologyAndEnum": "47th annual (1922)",
-                "customerCode": "PA",
-                "barcode": "32101065068437",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2083940,
-                "owningInstitutionItemId": "4862287",
-                "owningInstitutionHoldingsId": "5587786"
-              }
-            ],
-            "owningInstitutionBibId": "5436211",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 562145,
-            "title": "... Annual report to the ... Governor of the state of New Jersey / by the New Jersey Real Estate Commission.",
-            "author": " New Jersey Real Estate Commission.  ",
-            "publisher": "The Commission,",
-            "publisherDate": "1922-1923.",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1921/22-1922/23 -1938/39",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HD266.N4A5",
-                "chronologyAndEnum": "1921-/22-1938/39",
-                "customerCode": "PA",
-                "barcode": "32101047854086",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 682550,
-                "owningInstitutionItemId": "2892898",
-                "owningInstitutionHoldingsId": "2515913"
-              }
-            ],
-            "owningInstitutionBibId": "2228799",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1567417,
-            "title": "The ... biennial convention proceedings, District No. 22, United Mine Workers of America ...",
-            "author": " United Mine Workers of America. District 22.  ",
-            "publisher": "The District",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1st -3rd, (1914-1918);  Special, (1917);  5th-10th, (1922-1934);  12th-13th, (1939-1941).",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HD6515.M615 U52",
-                "chronologyAndEnum": "10th (1934)",
-                "customerCode": "PA",
-                "barcode": "32101077411781",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2209015,
-                "owningInstitutionItemId": "5859675",
-                "owningInstitutionHoldingsId": "6253051"
-              },
-              {
-                "callNumber": "HD6515.M615 U52",
-                "chronologyAndEnum": "12th (1939)",
-                "customerCode": "PA",
-                "barcode": "32101077411799",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2209016,
-                "owningInstitutionItemId": "5859681",
-                "owningInstitutionHoldingsId": "6253051"
-              },
-              {
-                "callNumber": "HD6515.M615 U52",
-                "chronologyAndEnum": "13th (1941)",
-                "customerCode": "PA",
-                "barcode": "32101077411807",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2209017,
-                "owningInstitutionItemId": "5859684",
-                "owningInstitutionHoldingsId": "6253051"
-              },
-              {
-                "callNumber": "HD6515.M615 U52",
-                "chronologyAndEnum": "1st (1914)",
-                "customerCode": "PA",
-                "barcode": "32101077411690",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2209006,
-                "owningInstitutionItemId": "5859638",
-                "owningInstitutionHoldingsId": "6253051"
-              },
-              {
-                "callNumber": "HD6515.M615 U52",
-                "chronologyAndEnum": "2nd (1916)",
-                "customerCode": "PA",
-                "barcode": "32101077411708",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2209007,
-                "owningInstitutionItemId": "5859639",
-                "owningInstitutionHoldingsId": "6253051"
-              },
-              {
-                "callNumber": "HD6515.M615 U52",
-                "chronologyAndEnum": "3rd (1918)",
-                "customerCode": "PA",
-                "barcode": "32101077411716",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2209008,
-                "owningInstitutionItemId": "5859640",
-                "owningInstitutionHoldingsId": "6253051"
-              },
-              {
-                "callNumber": "HD6515.M615 U52",
-                "chronologyAndEnum": "5th (1922)",
-                "customerCode": "PA",
-                "barcode": "32101077411732",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2209010,
-                "owningInstitutionItemId": "5859653",
-                "owningInstitutionHoldingsId": "6253051"
-              },
-              {
-                "callNumber": "HD6515.M615 U52",
-                "chronologyAndEnum": "6th (1924)",
-                "customerCode": "PA",
-                "barcode": "32101077411740",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2209011,
-                "owningInstitutionItemId": "5859657",
-                "owningInstitutionHoldingsId": "6253051"
-              },
-              {
-                "callNumber": "HD6515.M615 U52",
-                "chronologyAndEnum": "7th (1927)",
-                "customerCode": "PA",
-                "barcode": "32101077411757",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2209012,
-                "owningInstitutionItemId": "5859659",
-                "owningInstitutionHoldingsId": "6253051"
-              },
-              {
-                "callNumber": "HD6515.M615 U52",
-                "chronologyAndEnum": "8th (1930)",
-                "customerCode": "PA",
-                "barcode": "32101077411765",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2209013,
-                "owningInstitutionItemId": "5859662",
-                "owningInstitutionHoldingsId": "6253051"
-              },
-              {
-                "callNumber": "HD6515.M615 U52",
-                "chronologyAndEnum": "9th (1932)",
-                "customerCode": "PA",
-                "barcode": "32101077411773",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2209014,
-                "owningInstitutionItemId": "5859668",
-                "owningInstitutionHoldingsId": "6253051"
-              },
-              {
-                "callNumber": "HD6515.M615 U52",
-                "chronologyAndEnum": "Special (1917)",
-                "customerCode": "PA",
-                "barcode": "32101077411724",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2209009,
-                "owningInstitutionItemId": "5859645",
-                "owningInstitutionHoldingsId": "6253051"
-              }
-            ],
-            "owningInstitutionBibId": "6243253",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1776166,
-            "title": "... budget performance quarterly report / Malawi Government.",
-            "author": " Malawi, author.  ",
-            "publisher": "Ministry of Finance,",
-            "publisherDate": "[2010]-",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "2010/11 Q1-2010/11 Q2 ; 2011/12, Q1 (Oct. 2011)",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HJ83.2 .A23aq",
-                "chronologyAndEnum": "2010/11, Q1",
-                "customerCode": "PA",
-                "barcode": "32101092252780",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2668075,
-                "owningInstitutionItemId": "6986645",
-                "owningInstitutionHoldingsId": "8035071"
-              },
-              {
-                "callNumber": "HJ83.2 .A23aq",
-                "chronologyAndEnum": "2010/11, Q2",
-                "customerCode": "PA",
-                "barcode": "32101092252798",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2668076,
-                "owningInstitutionItemId": "6986647",
-                "owningInstitutionHoldingsId": "8035071"
-              },
-              {
-                "callNumber": "HJ83.2 .A23aq",
-                "chronologyAndEnum": "2011/12, Q1",
-                "customerCode": "PA",
-                "barcode": "32101091634319",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2668074,
-                "owningInstitutionItemId": "6865119",
-                "owningInstitutionHoldingsId": "8035071"
-              }
-            ],
-            "owningInstitutionBibId": "8282314",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 815991,
-            "title": "The ... Bulletin of the Central Intercollegiate Athletic Association ... with proceedings of the ... annual meeting.",
-            "author": " Central Intercollegiate Athletic Association.  ",
-            "publisher": "s.n.]",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1951-1953",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "4200.247",
-                "chronologyAndEnum": "1951-53",
-                "customerCode": "PA",
-                "barcode": "32101087116495",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1071773,
-                "owningInstitutionItemId": "6574332",
-                "owningInstitutionHoldingsId": "2878817"
-              }
-            ],
-            "owningInstitutionBibId": "2573981",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 598546,
-            "title": "... Compte géneral de l'administration des finances rendu ... ... Algemeene rekening van het bestuur van financien afgelegd ... door den Minister van financien ... par le Ministre des finances",
-            "author": " Belgium. Ministère des finances.  ",
-            "publisher": null,
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1913-1914, 1917-1918",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HJ53.A3q",
-                "chronologyAndEnum": "1913",
-                "customerCode": "PA",
-                "barcode": "32101078729942",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 728761,
-                "owningInstitutionItemId": "2634382",
-                "owningInstitutionHoldingsId": "2588153"
-              },
-              {
-                "callNumber": "HJ53.A3q",
-                "chronologyAndEnum": "1914",
-                "customerCode": "PA",
-                "barcode": "32101078729959",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 728760,
-                "owningInstitutionItemId": "2634379",
-                "owningInstitutionHoldingsId": "2588153"
-              },
-              {
-                "callNumber": "HJ53.A3q",
-                "chronologyAndEnum": "1917",
-                "customerCode": "PA",
-                "barcode": "32101078729603",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 728759,
-                "owningInstitutionItemId": "2634377",
-                "owningInstitutionHoldingsId": "2588153"
-              },
-              {
-                "callNumber": "HJ53.A3q",
-                "chronologyAndEnum": "1918",
-                "customerCode": "PA",
-                "barcode": "32101078729611",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 728758,
-                "owningInstitutionItemId": "2634374",
-                "owningInstitutionHoldingsId": "2588153"
-              }
-            ],
-            "owningInstitutionBibId": "2297446",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 8986342,
-            "title": "... energy, air and climate change.",
-            "author": "   ",
-            "publisher": "Bureau of Statistics",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "2014",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HD1698.L5 E647q",
-                "chronologyAndEnum": "2014",
-                "customerCode": "PA",
-                "barcode": "32101106513706",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14821518,
-                "owningInstitutionItemId": "7826769",
-                "owningInstitutionHoldingsId": "10516729"
-              }
-            ],
-            "owningInstitutionBibId": "10789983",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 688269,
-            "title": "... Financial abstract.",
-            "author": " London County Council.  ",
-            "publisher": null,
-            "publisherDate": "1897-19--",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1889/94, 1896/1906-1898/1908, 1910/11-1919/20",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HJ9041.L75",
-                "chronologyAndEnum": "1889/1904, 1898/1908",
-                "customerCode": "PA",
-                "barcode": "32101050298130",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 865028,
-                "owningInstitutionItemId": "2892416",
-                "owningInstitutionHoldingsId": "2659315"
-              }
-            ],
-            "owningInstitutionBibId": "2365456",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1287140,
-            "title": "... Grand Annual Entertainment & Dance of the Enosis Lefkariton of Cyprus in America, Inc. = Henōsis Leukaritōn en Amerikē : ... Megalē Etēsia Choroesperis.",
-            "author": "   ",
-            "publisher": "Enosis Lefkariton of Cyprus in America,",
-            "publisherDate": "1931-",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1968",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "E184.G7 G66",
-                "chronologyAndEnum": "1968",
-                "customerCode": "PA",
-                "barcode": "32101046403984",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1841119,
-                "owningInstitutionItemId": "3421922",
-                "owningInstitutionHoldingsId": "4192330"
-              }
-            ],
-            "owningInstitutionBibId": "3893122",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1829000,
-            "title": "... ICT statistics report.",
-            "author": "   ",
-            "publisher": "Statistics Botswana",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "2011",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HN806.Z9 I565q",
-                "chronologyAndEnum": "2011",
-                "customerCode": "PA",
-                "barcode": "32101098720459",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2731872,
-                "owningInstitutionItemId": "7377194",
-                "owningInstitutionHoldingsId": "9075149"
-              }
-            ],
-            "owningInstitutionBibId": "9199190",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1935257,
-            "title": "... International Aeronautical Conference.",
-            "author": "  International Aeronautical Conference. ",
-            "publisher": "The Institute of Aeronautical Sciences, Inc.,",
-            "publisherDate": "[1949].",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "2nd (24th-26th May 1949)",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "9250.114",
-                "chronologyAndEnum": "1949",
-                "customerCode": "PA",
-                "barcode": "32101045244355",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2944121,
-                "owningInstitutionItemId": "3925527",
-                "owningInstitutionHoldingsId": "782624"
-              }
-            ],
-            "owningInstitutionBibId": "701138",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1935259,
-            "title": "... International Aeronautical Conference.",
-            "author": "  International Aeronautical Conference. ",
-            "publisher": "Institute of the Aeronautical Sciences,",
-            "publisherDate": "[1955].",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "5th (1955)",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "9250.114",
-                "chronologyAndEnum": "1955",
-                "customerCode": "PA",
-                "barcode": "32101045244462",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2944126,
-                "owningInstitutionItemId": "3925544",
-                "owningInstitutionHoldingsId": "782644"
-              }
-            ],
-            "owningInstitutionBibId": "701158",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 9486403,
-            "title": "... İstanbul Film Festivali.",
-            "author": "  Uluslararası İstanbul Film Festivali. ",
-            "publisher": "[İstanbul Kültür Sanat Vakfı]",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "No. 31 (2012)",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "PN1993.4 .U48",
-                "chronologyAndEnum": "no.31 ((2012))",
-                "customerCode": "PA",
-                "barcode": "32101102550272",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 15752990,
-                "owningInstitutionItemId": "7849891",
-                "owningInstitutionHoldingsId": "10544814"
-              }
-            ],
-            "owningInstitutionBibId": "10824787",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1154321,
-            "title": "The ... Lawyer statistical report.",
-            "author": "   Lawyer statistical report (Chicago : 1967)",
-            "publisher": "American Bar Foundation,",
-            "publisherDate": "1968-1972.",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "7680.1275.2",
-                "chronologyAndEnum": "1967",
-                "customerCode": "PA",
-                "barcode": "32101043028644",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1604673,
-                "owningInstitutionItemId": "3228902",
-                "owningInstitutionHoldingsId": "392589"
-              },
-              {
-                "callNumber": "7680.1275.2",
-                "chronologyAndEnum": "1971",
-                "customerCode": "PA",
-                "barcode": "32101043028636",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1604674,
-                "owningInstitutionItemId": "3228903",
-                "owningInstitutionHoldingsId": "392589"
-              }
-            ],
-            "owningInstitutionBibId": "359288",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 9431218,
-            "title": "... Mazu wen hua nian jian.",
-            "author": "   ",
-            "publisher": "Fujian Sheng Lianjiang Xian zheng fu",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "2015",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "DS797.26.L537 M38q",
-                "chronologyAndEnum": "2015",
-                "customerCode": "PA",
-                "barcode": "32101098265760",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 15665497,
-                "owningInstitutionItemId": "8023910",
-                "owningInstitutionHoldingsId": "10867639"
-              }
-            ],
-            "owningInstitutionBibId": "11157665",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1336888,
-            "title": "... National Convention of the Pan-Macedonian Association.",
-            "author": " Pan-Macedonian Association. National Convention.  ",
-            "publisher": null,
-            "publisherDate": "Pan-Macedonian Association,",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "45th (1991)-46th (1992); 51st (1997)",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "E184.M3 P3",
-                "chronologyAndEnum": "45th (1991)",
-                "customerCode": "PA",
-                "barcode": "32101064372566",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1903806,
-                "owningInstitutionItemId": "5012429",
-                "owningInstitutionHoldingsId": "4475467"
-              },
-              {
-                "callNumber": "E184.M3 P3",
-                "chronologyAndEnum": "46th (1992)",
-                "customerCode": "PA",
-                "barcode": "32101064372558",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1903807,
-                "owningInstitutionItemId": "5012430",
-                "owningInstitutionHoldingsId": "4475467"
-              },
-              {
-                "callNumber": "E184.M3 P3",
-                "chronologyAndEnum": "51st (1997)",
-                "customerCode": "PA",
-                "barcode": "32101048256299",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1903805,
-                "owningInstitutionItemId": "3854338",
-                "owningInstitutionHoldingsId": "4475467"
-              }
-            ],
-            "owningInstitutionBibId": "4211563",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 9111722,
-            "title": "... programme based budget of the national Government of Kenya for the year ending ...",
-            "author": " Kenya, author.  ",
-            "publisher": "Government Printer",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "2017/18",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HJ81.3 A24f",
-                "chronologyAndEnum": "2017/18",
-                "customerCode": "PA",
-                "barcode": "32101102571377",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 15008840,
-                "owningInstitutionItemId": "7902295",
-                "owningInstitutionHoldingsId": "10664334"
-              }
-            ],
-            "owningInstitutionBibId": "10950982",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1875376,
-            "title": "... reconciliation report on the Petroleum Holding Fund / submitted to Parliament by ... Minister for Finance.",
-            "author": "   ",
-            "publisher": "[Ministry of Finance]",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "2014-2015",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HD9577.G4R43",
-                "chronologyAndEnum": "2014",
-                "customerCode": "PA",
-                "barcode": "32101098725367",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2789304,
-                "owningInstitutionItemId": "7547601",
-                "owningInstitutionHoldingsId": "9885651"
-              },
-              {
-                "callNumber": "HD9577.G4R43",
-                "chronologyAndEnum": "2015",
-                "customerCode": "PA",
-                "barcode": "32101099247817",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2789305,
-                "owningInstitutionItemId": "7547604",
-                "owningInstitutionHoldingsId": "9885651"
-              }
-            ],
-            "owningInstitutionBibId": "10091445",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 849002,
-            "title": "... Report of the Board of Directors of the Theological Seminary to the General Assembly of the Presbyterian Church.",
-            "author": " Princeton Theological Seminary.  ",
-            "publisher": null,
-            "publisherDate": "18uu-",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "66935.7379.999 v.2",
-                "chronologyAndEnum": "",
-                "customerCode": "PA",
-                "barcode": "32101068640943",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1021997,
-                "owningInstitutionItemId": "5044917",
-                "owningInstitutionHoldingsId": "2832013"
-              }
-            ],
-            "owningInstitutionBibId": "2626959",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 455136,
-            "title": "... Report of the director of emergency conservation work for the period ...",
-            "author": " Civilian Conservation Corps (U.S.)  ",
-            "publisher": "Civilian Conservation Corps :",
-            "publisherDate": "1934.",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1933-1934",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "9586.913",
-                "chronologyAndEnum": "1933-1933/1934",
-                "customerCode": "PA",
-                "barcode": "32101064431446",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 553901,
-                "owningInstitutionItemId": "4861907",
-                "owningInstitutionHoldingsId": "2344365"
-              },
-              {
-                "callNumber": "9586.913",
-                "chronologyAndEnum": "1934",
-                "customerCode": "PA",
-                "barcode": "32101064431453",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 553902,
-                "owningInstitutionItemId": "4861916",
-                "owningInstitutionHoldingsId": "2344365"
-              }
-            ],
-            "owningInstitutionBibId": "2061725",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1347283,
-            "title": "The ... report of the Joint National Committee on Detection, Evaluation, and Treatment of High Blood Pressure.",
-            "author": " Joint National Committee on Detection, Evaluation, and Treatment of High Blood Pressure.  ",
-            "publisher": "U.S. Dept. of Health and Human Services, Public Health Service, National Institutes of Health, National Heart, Lung, and Blood Institute, National High Blood Pressure Education Program,",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1984",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HE 20.3202:B 62/16/",
-                "chronologyAndEnum": "1984",
-                "customerCode": "PA",
-                "barcode": "32101057167882",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1918408,
-                "owningInstitutionItemId": "3932347",
-                "owningInstitutionHoldingsId": "4538453"
-              }
-            ],
-            "owningInstitutionBibId": "4280993",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 720794,
-            "title": "The ... Report of the New York City Housing Authority.",
-            "author": " New York City Housing Authority.  ",
-            "publisher": "The Authority,",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1954-1959",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HD7304.N4A52q",
-                "chronologyAndEnum": "1954-1962 (Inc.)",
-                "customerCode": "PA",
-                "barcode": "32101054890148",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 913873,
-                "owningInstitutionItemId": "4013113",
-                "owningInstitutionHoldingsId": "2703422"
-              }
-            ],
-            "owningInstitutionBibId": "2407775",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1467114,
-            "title": "... summer institute.",
-            "author": "   ",
-            "publisher": "Rutgers University",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "2nd, 1951",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "IR-7797",
-                "chronologyAndEnum": "2nd annual (1951)",
-                "customerCode": "PA",
-                "barcode": "32101065065482",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2081891,
-                "owningInstitutionItemId": "4850973",
-                "owningInstitutionHoldingsId": "5574271"
-              }
-            ],
-            "owningInstitutionBibId": "5420477",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1870934,
-            "title": "... Türkiye hapishaneler almanağı.",
-            "author": "   ",
-            "publisher": "Ceza İnfaz Sisteminde Sivil Toplum Derneği,",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "2015",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HV9776.7.T874",
-                "chronologyAndEnum": "2015",
-                "customerCode": "PA",
-                "barcode": "32101095029433",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2783742,
-                "owningInstitutionItemId": "7522933",
-                "owningInstitutionHoldingsId": "9820821"
-              }
-            ],
-            "owningInstitutionBibId": "10019222",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1115868,
-            "title": "0 yu 1 ke ji za zhi : Byte magazine Zhong wen ban = 0 & 1 technology : Byte magazine Chinese edition.",
-            "author": "   ",
-            "publisher": "Di san bo wen hua shi ye gu fen you xian gong si,",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "No.153 (1994)-no.166 (1995), no.169 (1995)-no.172 (1995), no.175 (1995)-no.188 (1996)",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.153-154 (1994)",
-                "customerCode": "PA",
-                "barcode": "32101049811662",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1505996,
-                "owningInstitutionItemId": "2503140",
-                "owningInstitutionHoldingsId": "3476656"
-              },
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.155-156 (1994)",
-                "customerCode": "PA",
-                "barcode": "32101049811670",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1505997,
-                "owningInstitutionItemId": "2503141",
-                "owningInstitutionHoldingsId": "3476656"
-              },
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.157-158 (1994)",
-                "customerCode": "PA",
-                "barcode": "32101049811688",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1505998,
-                "owningInstitutionItemId": "2503142",
-                "owningInstitutionHoldingsId": "3476656"
-              },
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.159-160 (1994)",
-                "customerCode": "PA",
-                "barcode": "32101049811696",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1505999,
-                "owningInstitutionItemId": "2503143",
-                "owningInstitutionHoldingsId": "3476656"
-              },
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.161-162 (1994)",
-                "customerCode": "PA",
-                "barcode": "32101049811704",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1506000,
-                "owningInstitutionItemId": "2503144",
-                "owningInstitutionHoldingsId": "3476656"
-              },
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.163-164 (1994)",
-                "customerCode": "PA",
-                "barcode": "32101049811712",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1506001,
-                "owningInstitutionItemId": "2503145",
-                "owningInstitutionHoldingsId": "3476656"
-              },
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.165-166 (1995)",
-                "customerCode": "PA",
-                "barcode": "32101049811720",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1506002,
-                "owningInstitutionItemId": "2503146",
-                "owningInstitutionHoldingsId": "3476656"
-              },
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.169-170 (1995)",
-                "customerCode": "PA",
-                "barcode": "32101049811738",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1506003,
-                "owningInstitutionItemId": "2503147",
-                "owningInstitutionHoldingsId": "3476656"
-              },
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.171-172 (1995)",
-                "customerCode": "PA",
-                "barcode": "32101049811746",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1506004,
-                "owningInstitutionItemId": "2503148",
-                "owningInstitutionHoldingsId": "3476656"
-              },
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.175-176 (1995)",
-                "customerCode": "PA",
-                "barcode": "32101049811753",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1506005,
-                "owningInstitutionItemId": "2503149",
-                "owningInstitutionHoldingsId": "3476656"
-              },
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.177-178 (1996)",
-                "customerCode": "PA",
-                "barcode": "32101049811761",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1506006,
-                "owningInstitutionItemId": "2503150",
-                "owningInstitutionHoldingsId": "3476656"
-              },
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.179-180 (1996)",
-                "customerCode": "PA",
-                "barcode": "32101049811779",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1506007,
-                "owningInstitutionItemId": "2503151",
-                "owningInstitutionHoldingsId": "3476656"
-              },
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.181-182 (1996)",
-                "customerCode": "PA",
-                "barcode": "32101049811787",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1506008,
-                "owningInstitutionItemId": "2503152",
-                "owningInstitutionHoldingsId": "3476656"
-              },
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.183-184 (1996)",
-                "customerCode": "PA",
-                "barcode": "32101049811795",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1506009,
-                "owningInstitutionItemId": "2503153",
-                "owningInstitutionHoldingsId": "3476656"
-              },
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.185-186 (1996)",
-                "customerCode": "PA",
-                "barcode": "32101049811803",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1506010,
-                "owningInstitutionItemId": "2503154",
-                "owningInstitutionHoldingsId": "3476656"
-              },
-              {
-                "callNumber": "QA75.5 .A12",
-                "chronologyAndEnum": "no.187-188 (1996)",
-                "customerCode": "PA",
-                "barcode": "32101049811811",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1506011,
-                "owningInstitutionItemId": "2503155",
-                "owningInstitutionHoldingsId": "3476656"
-              }
-            ],
-            "owningInstitutionBibId": "3162799",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1487377,
-            "title": "1/2 : nauchnye trudy, zapiski, issledovanii︠a︡.",
-            "author": "   ",
-            "publisher": "Morskoĭ gosudarstvennyĭ universitet imeni admirala G.I. Nevelʹskogo,",
-            "publisherDate": "2007-",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "No. 2 (Published in 2007)-no. 3 (published in 2008)",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HM791 .A1776",
-                "chronologyAndEnum": "no.2",
-                "customerCode": "PA",
-                "barcode": "32101061079909",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2107527,
-                "owningInstitutionItemId": "5176359",
-                "owningInstitutionHoldingsId": "5736708"
-              },
-              {
-                "callNumber": "HM791 .A1776",
-                "chronologyAndEnum": "no.3",
-                "customerCode": "PA",
-                "barcode": "32101073107839",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2107528,
-                "owningInstitutionItemId": "5430236",
-                "owningInstitutionHoldingsId": "5736708"
-              }
-            ],
-            "owningInstitutionBibId": "5571403",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1376240,
-            "title": "1 world manga.",
-            "author": "   ",
-            "publisher": "Viz ;",
-            "publisherDate": "c2006-",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "Passage 1- 6",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "PN6790.J3 A155",
-                "chronologyAndEnum": "Passage 1",
-                "customerCode": "PA",
-                "barcode": "32101057376541",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1955512,
-                "owningInstitutionItemId": "4176736",
-                "owningInstitutionHoldingsId": "4855558"
-              },
-              {
-                "callNumber": "PN6790.J3 A155",
-                "chronologyAndEnum": "Passage 2",
-                "customerCode": "PA",
-                "barcode": "32101061611966",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1955513,
-                "owningInstitutionItemId": "4341170",
-                "owningInstitutionHoldingsId": "4855558"
-              },
-              {
-                "callNumber": "PN6790.J3 A155",
-                "chronologyAndEnum": "Passage 3",
-                "customerCode": "PA",
-                "barcode": "32101061611503",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1955514,
-                "owningInstitutionItemId": "4341174",
-                "owningInstitutionHoldingsId": "4855558"
-              },
-              {
-                "callNumber": "PN6790.J3 A155",
-                "chronologyAndEnum": "Passage 4",
-                "customerCode": "PA",
-                "barcode": "32101063283830",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1955515,
-                "owningInstitutionItemId": "4384791",
-                "owningInstitutionHoldingsId": "4855558"
-              },
-              {
-                "callNumber": "PN6790.J3 A155",
-                "chronologyAndEnum": "Passage 5",
-                "customerCode": "PA",
-                "barcode": "32101065637306",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1955517,
-                "owningInstitutionItemId": "4740563",
-                "owningInstitutionHoldingsId": "4855558"
-              },
-              {
-                "callNumber": "PN6790.J3 A155",
-                "chronologyAndEnum": "Passage 6",
-                "customerCode": "PA",
-                "barcode": "32101060946488",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1955516,
-                "owningInstitutionItemId": "4580421",
-                "owningInstitutionHoldingsId": "4855558"
-              }
-            ],
-            "owningInstitutionBibId": "4618335",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1182281,
-            "title": "1 yılın kılavuzu, 1974 günler, haftalar, şenlikler, bayramlar, mesirler, ve fuarlar / Yaman Akbayar. [microform] :",
-            "author": "Akbayar, Yaman.   ",
-            "publisher": "[s.n.],",
-            "publisherDate": "1975-",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "[Nu.] I. 1974",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "DR401 .A33",
-                "chronologyAndEnum": "",
-                "customerCode": "PA",
-                "barcode": "32101099338244",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 1641718,
-                "owningInstitutionItemId": "7394775",
-                "owningInstitutionHoldingsId": "462035"
-              },
-              {
-                "callNumber": "MICROFILM 5504",
-                "chronologyAndEnum": "",
-                "customerCode": "PF",
-                "barcode": "32101030136939",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 744939,
-                "owningInstitutionItemId": "3244908",
-                "owningInstitutionHoldingsId": "462036"
-              },
-              {
-                "callNumber": "MICROFILM 11277",
-                "chronologyAndEnum": "",
-                "customerCode": "PF",
-                "barcode": "32101069842910",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 744940,
-                "owningInstitutionItemId": "6599299",
-                "owningInstitutionHoldingsId": "905134"
-              }
-            ],
-            "owningInstitutionBibId": "424319",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 9011301,
-            "title": "10 + 1.",
-            "author": "   ",
-            "publisher": "INAX,",
-            "publisherDate": "1994-2008.",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "No.1-no.50",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.1 (1994)",
-                "customerCode": "PA",
-                "barcode": "32101100199999",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862373,
-                "owningInstitutionItemId": "7857416",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.10 (1997)",
-                "customerCode": "PA",
-                "barcode": "32101100199502",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862352,
-                "owningInstitutionItemId": "7857434",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.11 (1997)",
-                "customerCode": "PA",
-                "barcode": "32101100199494",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862050,
-                "owningInstitutionItemId": "7857435",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.12 (1998)",
-                "customerCode": "PA",
-                "barcode": "32101100199486",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862042,
-                "owningInstitutionItemId": "7857436",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.13 (1998)",
-                "customerCode": "PA",
-                "barcode": "32101100199478",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862026,
-                "owningInstitutionItemId": "7857437",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.14 (1998)",
-                "customerCode": "PA",
-                "barcode": "32101100199460",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862003,
-                "owningInstitutionItemId": "7857438",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.15 (1998)",
-                "customerCode": "PA",
-                "barcode": "32101100199452",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862007,
-                "owningInstitutionItemId": "7857440",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.16 (1999)",
-                "customerCode": "PA",
-                "barcode": "32101100199445",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14861986,
-                "owningInstitutionItemId": "7857441",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.17 (1999)",
-                "customerCode": "PA",
-                "barcode": "32101100199437",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14861964,
-                "owningInstitutionItemId": "7857442",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.18 (1999)",
-                "customerCode": "PA",
-                "barcode": "32101100198751",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862730,
-                "owningInstitutionItemId": "7857903",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.19 (2000)",
-                "customerCode": "PA",
-                "barcode": "32101100198744",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862723,
-                "owningInstitutionItemId": "7857905",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.2 (1994)",
-                "customerCode": "PA",
-                "barcode": "32101100199981",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862353,
-                "owningInstitutionItemId": "7857417",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.20 (2000)",
-                "customerCode": "PA",
-                "barcode": "32101100198785",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862780,
-                "owningInstitutionItemId": "7857907",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.21 (2000)",
-                "customerCode": "PA",
-                "barcode": "32101100198777",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862765,
-                "owningInstitutionItemId": "7857909",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.22 (2000)",
-                "customerCode": "PA",
-                "barcode": "32101098300005",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862276,
-                "owningInstitutionItemId": "7857912",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.23 (2001)",
-                "customerCode": "PA",
-                "barcode": "32101098300070",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862200,
-                "owningInstitutionItemId": "7857914",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.24 (2001)",
-                "customerCode": "PA",
-                "barcode": "32101098300062",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862181,
-                "owningInstitutionItemId": "7857924",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.25 (2001)",
-                "customerCode": "PA",
-                "barcode": "32101105195281",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862923,
-                "owningInstitutionItemId": "7808436",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.26 (2002)",
-                "customerCode": "PA",
-                "barcode": "32101098300054",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862223,
-                "owningInstitutionItemId": "7857927",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.27 (2002)",
-                "customerCode": "PA",
-                "barcode": "32101098300047",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862203,
-                "owningInstitutionItemId": "7857929",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.28 (2002)",
-                "customerCode": "PA",
-                "barcode": "32101098300039",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862250,
-                "owningInstitutionItemId": "7857931",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.29 (2002)",
-                "customerCode": "PA",
-                "barcode": "32101098300021",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862244,
-                "owningInstitutionItemId": "7857935",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.3 (1995)",
-                "customerCode": "PA",
-                "barcode": "32101100199916",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862266,
-                "owningInstitutionItemId": "7857418",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.30 (2003)",
-                "customerCode": "PA",
-                "barcode": "32101098300013",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862296,
-                "owningInstitutionItemId": "7857938",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.31 (2003)",
-                "customerCode": "PA",
-                "barcode": "32101098300260",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14863141,
-                "owningInstitutionItemId": "7857944",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.32 (2003)",
-                "customerCode": "PA",
-                "barcode": "32101098300252",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14863202,
-                "owningInstitutionItemId": "7857946",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.33 (2003)",
-                "customerCode": "PA",
-                "barcode": "32101098300245",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14863194,
-                "owningInstitutionItemId": "7857951",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.34 (2004)",
-                "customerCode": "PA",
-                "barcode": "32101098300237",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862985,
-                "owningInstitutionItemId": "7857952",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.35 (2004)",
-                "customerCode": "PA",
-                "barcode": "32101098300229",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862993,
-                "owningInstitutionItemId": "7857954",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.36 (2004)",
-                "customerCode": "PA",
-                "barcode": "32101098300211",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14863014,
-                "owningInstitutionItemId": "7857956",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.37 (2004)",
-                "customerCode": "PA",
-                "barcode": "32101098300203",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14863015,
-                "owningInstitutionItemId": "7857957",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.38 (2005)",
-                "customerCode": "PA",
-                "barcode": "32101098300195",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14861732,
-                "owningInstitutionItemId": "7857960",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.39 (2005)",
-                "customerCode": "PA",
-                "barcode": "32101098300187",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14861705,
-                "owningInstitutionItemId": "7857963",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.4 (1995)",
-                "customerCode": "PA",
-                "barcode": "32101100199924",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862273,
-                "owningInstitutionItemId": "7857421",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.40 (2005)",
-                "customerCode": "PA",
-                "barcode": "32101098300179",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14861769,
-                "owningInstitutionItemId": "7857974",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.41 (2005)",
-                "customerCode": "PA",
-                "barcode": "32101100198769",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862748,
-                "owningInstitutionItemId": "7857977",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.42 (2006)",
-                "customerCode": "PA",
-                "barcode": "32101098300161",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14861744,
-                "owningInstitutionItemId": "7857979",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.43 (2006)",
-                "customerCode": "PA",
-                "barcode": "32101098300153",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14861783,
-                "owningInstitutionItemId": "7857981",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.44 (2006)",
-                "customerCode": "PA",
-                "barcode": "32101098300146",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14861786,
-                "owningInstitutionItemId": "7857983",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.45 (2006)",
-                "customerCode": "PA",
-                "barcode": "32101098300138",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14861806,
-                "owningInstitutionItemId": "7857985",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.46 (2007)",
-                "customerCode": "PA",
-                "barcode": "32101098300120",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14861803,
-                "owningInstitutionItemId": "7857988",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.47 (2007)",
-                "customerCode": "PA",
-                "barcode": "32101098300112",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14861605,
-                "owningInstitutionItemId": "7857991",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.48 (2007)",
-                "customerCode": "PA",
-                "barcode": "32101098300088",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862367,
-                "owningInstitutionItemId": "7857996",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.49 (2007)",
-                "customerCode": "PA",
-                "barcode": "32101098300104",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14861596,
-                "owningInstitutionItemId": "7858007",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.5 (1996)",
-                "customerCode": "PA",
-                "barcode": "32101100199932",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862297,
-                "owningInstitutionItemId": "7857425",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.50 (2008)",
-                "customerCode": "PA",
-                "barcode": "32101098300096",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862394,
-                "owningInstitutionItemId": "7858009",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.6 (1996)",
-                "customerCode": "PA",
-                "barcode": "32101100199940",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862304,
-                "owningInstitutionItemId": "7857427",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.7 (1996)",
-                "customerCode": "PA",
-                "barcode": "32101100199957",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862336,
-                "owningInstitutionItemId": "7857430",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.8 (1997)",
-                "customerCode": "PA",
-                "barcode": "32101100199973",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862347,
-                "owningInstitutionItemId": "7857431",
-                "owningInstitutionHoldingsId": "10267100"
-              },
-              {
-                "callNumber": "NA6.J33 A13",
-                "chronologyAndEnum": "no.9 (1997)",
-                "customerCode": "PA",
-                "barcode": "32101100199965",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 14862339,
-                "owningInstitutionItemId": "7857433",
-                "owningInstitutionHoldingsId": "10267100"
-              }
-            ],
-            "owningInstitutionBibId": "10516387",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1845407,
-            "title": "10-K reports [microform].",
-            "author": "   ",
-            "publisher": "Disclosure",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1966-1987",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "A-Ainslie (box 1)",
-                "customerCode": "PF",
-                "barcode": "32101082932011",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752627,
-                "owningInstitutionItemId": "7340072",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Air-Altron (box 2)",
-                "customerCode": "PF",
-                "barcode": "32101082932193",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752628,
-                "owningInstitutionItemId": "7340073",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Alumacraft-American Guaranty (box 3)",
-                "customerCode": "PF",
-                "barcode": "32101082932029",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752629,
-                "owningInstitutionItemId": "7340074",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "American Home Assurance-American Syn Fuels (box 4)",
-                "customerCode": "PF",
-                "barcode": "32101082932185",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752630,
-                "owningInstitutionItemId": "7340075",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "American Technical-Apple Computer (box 5)",
-                "customerCode": "PF",
-                "barcode": "32101082932177",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752631,
-                "owningInstitutionItemId": "7340078",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Applebaums-Atlantic (box 6)",
-                "customerCode": "PF",
-                "barcode": "32101082932037",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752632,
-                "owningInstitutionItemId": "7340079",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Atlantis-Bank Corporation (box 7)",
-                "customerCode": "PF",
-                "barcode": "32101082932045",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752633,
-                "owningInstitutionItemId": "7340080",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Bank of America-Bemis (box 8)",
-                "customerCode": "PF",
-                "barcode": "32101082932052",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752634,
-                "owningInstitutionItemId": "7340081",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Ben-Boothe (box 9)",
-                "customerCode": "PF",
-                "barcode": "32101082932201",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752635,
-                "owningInstitutionItemId": "7340088",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Borden-Bunting (box 10)",
-                "customerCode": "PF",
-                "barcode": "32101082932060",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752636,
-                "owningInstitutionItemId": "7340090",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Burdett-Camseal (box 11)",
-                "customerCode": "PF",
-                "barcode": "32101082932078",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752637,
-                "owningInstitutionItemId": "7340091",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "CRA-Damson (box 19)",
-                "customerCode": "PF",
-                "barcode": "32101082932151",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752645,
-                "owningInstitutionItemId": "7341012",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Can Am-Casualtiy Insurance (box 12)",
-                "customerCode": "PF",
-                "barcode": "32101082932086",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752638,
-                "owningInstitutionItemId": "7340092",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Catalina-Cezar (box 13)",
-                "customerCode": "PF",
-                "barcode": "32101082932094",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752639,
-                "owningInstitutionItemId": "7340098",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Champion-Chrysler (box 14)",
-                "customerCode": "PF",
-                "barcode": "32101082932102",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752640,
-                "owningInstitutionItemId": "7340099",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Chubb-Coit (box 15)",
-                "customerCode": "PF",
-                "barcode": "32101082932110",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752641,
-                "owningInstitutionItemId": "7340100",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Coleco-Communications (box 16)",
-                "customerCode": "PF",
-                "barcode": "32101082932128",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752642,
-                "owningInstitutionItemId": "7340101",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Community-Consolidated (box 17)",
-                "customerCode": "PF",
-                "barcode": "32101082932136",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752643,
-                "owningInstitutionItemId": "7341007",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Constellation-CPU (box 18)",
-                "customerCode": "PF",
-                "barcode": "32101082932144",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752644,
-                "owningInstitutionItemId": "7341009",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Dan River-Desert (box 20)",
-                "customerCode": "PF",
-                "barcode": "32101082937382",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752648,
-                "owningInstitutionItemId": "7341040",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Design-Dresser (box 21)",
-                "customerCode": "PF",
-                "barcode": "32101082937374",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752649,
-                "owningInstitutionItemId": "7341043",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Drever-Educator (box 22)",
-                "customerCode": "PF",
-                "barcode": "32101082937366",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752650,
-                "owningInstitutionItemId": "7341044",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Educators-Energy North (box 23)",
-                "customerCode": "PF",
-                "barcode": "32101082932169",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752646,
-                "owningInstitutionItemId": "7341027",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Energy Search-Extractive (box 24)",
-                "customerCode": "PF",
-                "barcode": "32101082937390",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752647,
-                "owningInstitutionItemId": "7341030",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Eyecare-Firestone (box 25)",
-                "customerCode": "PF",
-                "barcode": "32101082936814",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752651,
-                "owningInstitutionItemId": "7349275",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "First A-First Virginia (box 26)",
-                "customerCode": "PF",
-                "barcode": "32101082936822",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752652,
-                "owningInstitutionItemId": "7349280",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "First Wachovia-Frawley (box 27)",
-                "customerCode": "PF",
-                "barcode": "32101082936830",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752653,
-                "owningInstitutionItemId": "7349284",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Fred-General Growth (box 28)",
-                "customerCode": "PF",
-                "barcode": "32101082936848",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752654,
-                "owningInstitutionItemId": "7349310",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "GNAC-GRT (box 30)",
-                "customerCode": "PF",
-                "barcode": "32101082936863",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752656,
-                "owningInstitutionItemId": "7349322",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "General Health-GMR (box 29)",
-                "customerCode": "PF",
-                "barcode": "32101082936855",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752655,
-                "owningInstitutionItemId": "7349320",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Grubb-Herzing (box 31)",
-                "customerCode": "PF",
-                "barcode": "32101082936871",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752657,
-                "owningInstitutionItemId": "7349368",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Hess-Holiday Resources (box 32)",
-                "customerCode": "PF",
-                "barcode": "32101082936889",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752658,
-                "owningInstitutionItemId": "7349374",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Holiday Universal-ICX (box 33)",
-                "customerCode": "PF",
-                "barcode": "32101082936897",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752659,
-                "owningInstitutionItemId": "7349381",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "ID-Interbanc (box 34)",
-                "customerCode": "PF",
-                "barcode": "32101082936905",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752660,
-                "owningInstitutionItemId": "7349450",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "IREX-Joslyn (box 36)",
-                "customerCode": "PF",
-                "barcode": "32101082936921",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752662,
-                "owningInstitutionItemId": "7349459",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Intercare-IPM (box 35)",
-                "customerCode": "PF",
-                "barcode": "32101082936913",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752661,
-                "owningInstitutionItemId": "7349454",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Joslyn (2)-King (box 37)",
-                "customerCode": "PF",
-                "barcode": "32101082936939",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752663,
-                "owningInstitutionItemId": "7354035",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Kings-Leroy (box 38)",
-                "customerCode": "PF",
-                "barcode": "32101082936947",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752664,
-                "owningInstitutionItemId": "7354038",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Lescarden-Ludlow (box 39)",
-                "customerCode": "PF",
-                "barcode": "32101082936954",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752665,
-                "owningInstitutionItemId": "7354041",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Lufkin-Martin Marieta (box 40)",
-                "customerCode": "PF",
-                "barcode": "32101082936962",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752666,
-                "owningInstitutionItemId": "7354057",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Martin Processing-Medical Care (box 41)",
-                "customerCode": "PF",
-                "barcode": "32101082936970",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752667,
-                "owningInstitutionItemId": "7354062",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Medical Comm.-Middlesex Water (box 42)",
-                "customerCode": "PF",
-                "barcode": "32101082936988",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752668,
-                "owningInstitutionItemId": "7354064",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Midland-Monfort (box 43)",
-                "customerCode": "PF",
-                "barcode": "32101082936996",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752669,
-                "owningInstitutionItemId": "7354074",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Monica-National By Products (box 44)",
-                "customerCode": "PF",
-                "barcode": "32101082937002",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752670,
-                "owningInstitutionItemId": "7354075",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "National Can-New England Life Pen (box 45)",
-                "customerCode": "PF",
-                "barcode": "32101082937010",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752671,
-                "owningInstitutionItemId": "7354094",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "New England Life Pension-Northern Arizona (box 46)",
-                "customerCode": "PF",
-                "barcode": "32101082937028",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752672,
-                "owningInstitutionItemId": "7354097",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Northern California-OGR (box 47)",
-                "customerCode": "PF",
-                "barcode": "32101082937036",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752673,
-                "owningInstitutionItemId": "7354099",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Ohio-Pacific Express (box 48)",
-                "customerCode": "PF",
-                "barcode": "32101082937044",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752674,
-                "owningInstitutionItemId": "7354101",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Pacific Far-Pay N Pak (box 49)",
-                "customerCode": "PF",
-                "barcode": "32101082937051",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752675,
-                "owningInstitutionItemId": "7354133",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Pay N Save-Pfizer (box 50)",
-                "customerCode": "PF",
-                "barcode": "32101082937069",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752676,
-                "owningInstitutionItemId": "7354135",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Pharmacaps-Porex (box 51)",
-                "customerCode": "PF",
-                "barcode": "32101082937077",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752677,
-                "owningInstitutionItemId": "7354149",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Port Huron-Purex (box 52)",
-                "customerCode": "PF",
-                "barcode": "32101082937085",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752678,
-                "owningInstitutionItemId": "7354150",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Purification-Recreama (box 53)",
-                "customerCode": "PF",
-                "barcode": "32101082937093",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752679,
-                "owningInstitutionItemId": "7354160",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Recreational-Roberts Oil (box 54)",
-                "customerCode": "PF",
-                "barcode": "32101082937101",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752680,
-                "owningInstitutionItemId": "7354161",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Robertshaw-San Bar (box 55)",
-                "customerCode": "PF",
-                "barcode": "32101082937119",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752681,
-                "owningInstitutionItemId": "7354162",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "San Carlos-SEBRN (box 56)",
-                "customerCode": "PF",
-                "barcode": "32101082937127",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752682,
-                "owningInstitutionItemId": "7354165",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Secom-Silvey (box 57)",
-                "customerCode": "PF",
-                "barcode": "32101082937135",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752683,
-                "owningInstitutionItemId": "7354472",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Simkar-Southwest Airlines (box 58)",
-                "customerCode": "PF",
-                "barcode": "32101082937143",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752684,
-                "owningInstitutionItemId": "7354473",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Southwest Airmotive-Stapling (box 59)",
-                "customerCode": "PF",
-                "barcode": "32101082937150",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752685,
-                "owningInstitutionItemId": "7354476",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Star-Super 8 Economy (box 60)",
-                "customerCode": "PF",
-                "barcode": "32101082937168",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752686,
-                "owningInstitutionItemId": "7354508",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Super 8 Lodging-Telectro (box 61)",
-                "customerCode": "PF",
-                "barcode": "32101082937176",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752687,
-                "owningInstitutionItemId": "7354509",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "TRE-Unioil (box 64)",
-                "customerCode": "PF",
-                "barcode": "32101082937200",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752690,
-                "owningInstitutionItemId": "7354514",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Teledata-Tiffany (box 62)",
-                "customerCode": "PF",
-                "barcode": "32101082937184",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752688,
-                "owningInstitutionItemId": "7354510",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Tiger-TRC (box 63)",
-                "customerCode": "PF",
-                "barcode": "32101082937192",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752689,
-                "owningInstitutionItemId": "7354512",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Union-United West (box 65)",
-                "customerCode": "PF",
-                "barcode": "32101082937218",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752691,
-                "owningInstitutionItemId": "7354515",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Unitek-Vidtor (box 66)",
-                "customerCode": "PF",
-                "barcode": "32101082937226",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752692,
-                "owningInstitutionItemId": "7354516",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Vie-Weingarten Properties (box 67)",
-                "customerCode": "PF",
-                "barcode": "32101082937358",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752693,
-                "owningInstitutionItemId": "7355841",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Weingarten Realty-Wilmington (box 68)",
-                "customerCode": "PF",
-                "barcode": "32101082937341",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752694,
-                "owningInstitutionItemId": "7355842",
-                "owningInstitutionHoldingsId": "9416703"
-              },
-              {
-                "callNumber": "MICROFICHE",
-                "chronologyAndEnum": "Wilsl-745 (box 69)",
-                "customerCode": "PF",
-                "barcode": "32101082937234",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2752695,
-                "owningInstitutionItemId": "7355843",
-                "owningInstitutionHoldingsId": "9416703"
-              }
-            ],
-            "owningInstitutionBibId": "9569951",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1904346,
-            "title": "The 100 best investments.",
-            "author": "   ",
-            "publisher": "The British, Foreign and Colonial Corporation, Ltd.,",
-            "publisherDate": null,
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "9th (Aug. 1919)-15th (1925)",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "HG4502.O58",
-                "chronologyAndEnum": "1920",
-                "customerCode": "PA",
-                "barcode": "32101073867150",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2880255,
-                "owningInstitutionItemId": "5349924",
-                "owningInstitutionHoldingsId": "691678"
-              },
-              {
-                "callNumber": "HG4502.O58",
-                "chronologyAndEnum": "1921/22",
-                "customerCode": "PA",
-                "barcode": "32101073867168",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2880256,
-                "owningInstitutionItemId": "5349927",
-                "owningInstitutionHoldingsId": "691678"
-              },
-              {
-                "callNumber": "HG4502.O58",
-                "chronologyAndEnum": "1922/23",
-                "customerCode": "PA",
-                "barcode": "32101073867176",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2880257,
-                "owningInstitutionItemId": "5349930",
-                "owningInstitutionHoldingsId": "691678"
-              },
-              {
-                "callNumber": "HG4502.O58",
-                "chronologyAndEnum": "1923/24",
-                "customerCode": "PA",
-                "barcode": "32101073867184",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2880258,
-                "owningInstitutionItemId": "5349934",
-                "owningInstitutionHoldingsId": "691678"
-              },
-              {
-                "callNumber": "HG4502.O58",
-                "chronologyAndEnum": "1924",
-                "customerCode": "PA",
-                "barcode": "32101073867192",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2880259,
-                "owningInstitutionItemId": "5349941",
-                "owningInstitutionHoldingsId": "691678"
-              },
-              {
-                "callNumber": "HG4502.O58",
-                "chronologyAndEnum": "1925",
-                "customerCode": "PA",
-                "barcode": "32101073867200",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2880260,
-                "owningInstitutionItemId": "5349943",
-                "owningInstitutionHoldingsId": "691678"
-              },
-              {
-                "callNumber": "HG4502.O58",
-                "chronologyAndEnum": "Aug.1919",
-                "customerCode": "PA",
-                "barcode": "32101073867143",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2880254,
-                "owningInstitutionItemId": "5349922",
-                "owningInstitutionHoldingsId": "691678"
-              }
-            ],
-            "owningInstitutionBibId": "629736",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 5,
-            "title": "100 voprosov - 100 otvetov.",
-            "author": "   ",
-            "publisher": "Mol. Gvardii︠a︡,",
-            "publisherDate": "1972-",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "1977, 1979, 1981-1982, 1985-1986",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "UA772 .S82",
-                "chronologyAndEnum": "1977 (vyp.6)",
-                "customerCode": "PA",
-                "barcode": "32101075649135",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 17,
-                "owningInstitutionItemId": "5591105",
-                "owningInstitutionHoldingsId": "23"
-              },
-              {
-                "callNumber": "UA772 .S82",
-                "chronologyAndEnum": "1979 (vyp.7)",
-                "customerCode": "PA",
-                "barcode": "32101075649143",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 18,
-                "owningInstitutionItemId": "5591314",
-                "owningInstitutionHoldingsId": "23"
-              },
-              {
-                "callNumber": "UA772 .S82",
-                "chronologyAndEnum": "1981 (vyp.8)",
-                "customerCode": "PA",
-                "barcode": "32101075649150",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 19,
-                "owningInstitutionItemId": "5591315",
-                "owningInstitutionHoldingsId": "23"
-              },
-              {
-                "callNumber": "UA772 .S82",
-                "chronologyAndEnum": "1982 (vyp.8/2)",
-                "customerCode": "PA",
-                "barcode": "32101075649168",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 20,
-                "owningInstitutionItemId": "5591317",
-                "owningInstitutionHoldingsId": "23"
-              },
-              {
-                "callNumber": "UA772 .S82",
-                "chronologyAndEnum": "1985",
-                "customerCode": "PA",
-                "barcode": "32101075649176",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 21,
-                "owningInstitutionItemId": "5591320",
-                "owningInstitutionHoldingsId": "23"
-              },
-              {
-                "callNumber": "UA772 .S82",
-                "chronologyAndEnum": "1986",
-                "customerCode": "PA",
-                "barcode": "32101075649184",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 22,
-                "owningInstitutionItemId": "5591324",
-                "owningInstitutionHoldingsId": "23"
-              }
-            ],
-            "owningInstitutionBibId": "22",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1528045,
-            "title": "11. Copenhagen Open, 18.-28. Juni 1989.",
-            "author": "   ",
-            "publisher": "Københavns skak union,",
-            "publisherDate": "1989.",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "Bulletin nr. 1-bulletin nr. 10",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "GV1313 .C66",
-                "chronologyAndEnum": "bulletin nr. 1-10",
-                "customerCode": "PA",
-                "barcode": "32101053685861",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2157093,
-                "owningInstitutionItemId": "5545569",
-                "owningInstitutionHoldingsId": "6006452"
-              }
-            ],
-            "owningInstitutionBibId": "5940152",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 2079954,
-            "title": "1199 news / 1199 National Health and Human Service Employees Union.",
-            "author": "   1199 news (1199, National Health and Human Service Employees Union)",
-            "publisher": "The Union,",
-            "publisherDate": "1993-",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "Vol. 13, no. 7-8 (July/Aug. 1995)-v. 17, no. 10 (Dec. 1999)",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "RA971.35.A199q",
-                "chronologyAndEnum": "vol. 13, no. 7/8-v. 15 (inc.) (1995-1997)",
-                "customerCode": "PA",
-                "barcode": "32101043351871",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 3176292,
-                "owningInstitutionItemId": "2640042",
-                "owningInstitutionHoldingsId": "3965728"
-              },
-              {
-                "callNumber": "RA971.35.A199q",
-                "chronologyAndEnum": "vol. 16-17 (1998-1999)",
-                "customerCode": "PA",
-                "barcode": "32101043351863",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 3176293,
-                "owningInstitutionItemId": "2640043",
-                "owningInstitutionHoldingsId": "3965728"
-              }
-            ],
-            "owningInstitutionBibId": "1051794",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1528056,
-            "title": "12. Copenhagen Open, 20.-30. Juni 1990.",
-            "author": "   ",
-            "publisher": "Københavns skak union,",
-            "publisherDate": "1990.",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "Bulletin nr. 1-bulletin nr. 10 ; introd. issue",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "GV1313 .C662",
-                "chronologyAndEnum": "bulletin nr. 1-10; introd. issue",
-                "customerCode": "PA",
-                "barcode": "32101053685879",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2157104,
-                "owningInstitutionItemId": "5545594",
-                "owningInstitutionHoldingsId": "6006459"
-              }
-            ],
-            "owningInstitutionBibId": "5940177",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          },
-          {
-            "bibId": 1528057,
-            "title": "13. Copenhagen Open, 24. Juni- 4. Juli 1991.",
-            "author": "   ",
-            "publisher": "Københavns skak union,",
-            "publisherDate": "1991.",
-            "owningInstitution": "PUL",
-            "customerCode": null,
-            "collectionGroupDesignation": null,
-            "useRestriction": null,
-            "barcode": null,
-            "summaryHoldings": "Bulletin nr. 1-bulletin nr. 10",
-            "availability": "Available",
-            "leaderMaterialType": "Serial",
-            "selected": false,
-            "showItems": true,
-            "selectAllItems": false,
-            "itemId": null,
-            "searchItemResultRows": [
-              {
-                "callNumber": "GV1313 .C663",
-                "chronologyAndEnum": "bulletin nr. 1-10",
-                "customerCode": "PA",
-                "barcode": "32101053685887",
-                "useRestriction": "No Restrictions",
-                "collectionGroupDesignation": "Shared",
-                "availability": "Available",
-                "selectedItem": false,
-                "itemId": 2157105,
-                "owningInstitutionItemId": "5545621",
-                "owningInstitutionHoldingsId": "6006465"
-              }
-            ],
-            "owningInstitutionBibId": "5940189",
-            "owningInstitutionHoldingsId": null,
-            "owningInstitutionItemId": null,
-            "requestPosition": null,
-            "patronBarcode": null,
-            "requestingInstitution": null,
-            "deliveryLocation": null,
-            "requestType": null,
-            "requestNotes": null
-          }
-      
-];
