--- src/app/modules/home/data.ts
@@ -1,124 +1,0 @@
-export const EMPLOYEE_DATA = [
-    {
-        Age: 55,
-        Employees: [
-            {
-                Age: 43,
-                Employees: [],
-                HireDate: new Date(2011, 6, 3),
-                ID: 3,
-                Name: "Michael Burke"
-            },
-            {
-                Age: 29,
-                Employees: undefined,
-                HireDate: new Date(2009, 6, 19),
-                ID: 2,
-                Name: "Thomas Anderson"
-            },
-            {
-                Age: 31,
-                Employees: [
-                    {
-                        Age: 35,
-                        HireDate: new Date(2015, 9, 17),
-                        ID: 6,
-                        Name: "Roland Mendel"
-                    },
-                    {
-                        Age: 44,
-                        HireDate: new Date(2009, 10, 11),
-                        ID: 12,
-                        Name: "Sven Cooper"
-                    }
-                ],
-                HireDate: new Date(2014, 8, 18),
-                ID: 11,
-                Name: "Monica Reyes"
-            }],
-        HireDate: new Date(2008, 3, 20),
-        ID: 1,
-        Name: "Johnathan Winchester"
-    },
-    {
-        Age: 42,
-        Employees: [
-            {
-                Age: 44,
-                HireDate: new Date(2014, 4, 4),
-                ID: 14,
-                Name: "Laurence Johnson"
-            },
-            {
-                Age: 25,
-                Employees: [
-                    {
-                        Age: 39,
-                        HireDate: new Date(2010, 3, 22),
-                        ID: 13,
-                        Name: "Trevor Ashworth"
-                    }
-                ],
-                HireDate: new Date(2017, 11, 9),
-                ID: 5,
-                Name: "Elizabeth Richards"
-            }],
-        HireDate: new Date(2014, 1, 22),
-        ID: 4,
-        Name: "Ana Sanders"
-    },
-    {
-        Age: 49,
-        Employees: [
-            {
-                Age: 44,
-                Employees: [],
-                HireDate: new Date(2014, 4, 4),
-                ID: 17,
-                Name: "Antonio Moreno"
-            }],
-        HireDate: new Date(2014, 1, 22),
-        ID: 18,
-        Name: "Victoria Lincoln"
-    },
-    {
-        Age: 61,
-        Employees: [
-            {
-                Age: 50,
-                Employees: [
-                    {
-                        Age: 25,
-                        Employees: [],
-                        HireDate: new Date(2017, 11, 9),
-                        ID: 15,
-                        Name: "Patricia Simpson"
-                    },
-                    {
-                        Age: 39,
-                        HireDate: new Date(2010, 3, 22),
-                        ID: 9,
-                        Name: "Francisco Chang"
-                    },
-                    {
-                        Age: 25,
-                        HireDate: new Date(2018, 3, 18),
-                        ID: 16,
-                        Name: "Peter Lewis"
-                    }
-                ],
-                HireDate: new Date(2007, 11, 18),
-                ID: 7,
-                Name: "Pedro Rodriguez"
-            },
-            {
-                Age: 27,
-                HireDate: new Date(2016, 2, 19),
-                ID: 8,
-                Name: "Casey Harper"
-            }],
-        HireDate: new Date(2010, 1, 1),
-        ID: 10,
-        Name: "Yang Wang"
-    }
-];
